# docs/openapi-schema.yml
openapi: 3.0.2
info:
  title: 'Foodgram'
  version: ''
paths:
...
  /api/recipes/{id}/:
    get:
      operationId: Get recipe
      description: ''
      parameters:
        - name: id
          in: path
          required: true
          description: "Unique user identifier"
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeList'
          description: ''
      tags:
        - Recipe
    patch:
      operationId: Update recipe
      security:
        - Token: [ ]
      description: 'Access only for creator of recipe'
      parameters:
        - name: id
          in: path
          required: true
          description: "Unique recipe identifier."
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeCreateUpdate'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeList'
          description: "Recipe successfully created"
        '400':
          $ref: '#/components/responses/NestedValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Recipes
    delete:
      operationId: Delete recipe

      description: 'Access only for creator of recipe'
      security:
        - Token: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: "Unique recipe identifier."
          schema:
            type: string
      responses:
        '204':
          description: "Recipe successfully created"
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Recipes
...
components:
  schemas:
    User:
      description: 'User (Author in recipe)'
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: "Email"
        id:
          type: integer
          readOnly: true
        username:
          type: string
          description: "Unique username"
          pattern: ^[\w.@+-]+\z
          maxLength: 150
        first_name:
          type: string
          maxLength: 150
          description: "Name"
          example: "John"
        last_name:
          type: string
          maxLength: 150
          description: "Surname"
          example: "Doe"
        is_subscribed:
          type: boolean
          readOnly: true
          description: "Is current user subscribed on pointed"
          example: false
      required:
        - username
...
    Tag:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
          description: 'Title'
          example: 'Breakfast'
        color:
          type: string
          nullable: true
          maxLength: 7
          description: 'Color in HEX'
          example: '#E26C2D'
        slug:
          type: string
          nullable: true
          maxLength: 200
          pattern: ^[-a-zA-Z0-9_]+$
          description: 'Unique slug'
          example: 'breakfast'
    RecipeList:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: 'Unique id'
        tags:
          description: 'Tag list'
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        author:
          $ref: '#/components/schemas/User'
        ingredients:
          description: 'Ingredient list'
          type: array
          items:
            $ref: '#/components/schemas/IngredientInRecipe'
        is_favorited:
          type: boolean
          description: 'Is in favorites'
        is_in_shopping_cart:
          type: boolean
          description: 'Is in basket'
        name:
          type: string
          maxLength: 200
          description: 'Title'
        image:
          description: 'Link on image on site'
          example: 'http://foodgram.example.org/media/recipes/images/image.jpeg'
          type: string
          format: url
        text:
          description: 'Description'
          type: string
        cooking_time:
          description: 'Cooking time (in minutes)'
          type: integer
          minimum: 1
      required:
        - tags
        - author
        - is_favorited
        - is_in_shopping_cart
        - name
        - image
        - text
        - cooking_time
...
    IngredientInRecipe:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
          description: 'Title'
          example: 'Boiled potatoes'
        measurement_unit:
          type: string
          maxLength: 200
          description: 'Units'
          example: 'g'
        amount:
          type: integer
          description: 'Amount'
          minimum: 1

      required:
        - name
        - measurement_unit
...
    RecipeCreateUpdate:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        ingredients:
          description: Ingredient list
          type: array
          items:
            example:
              id: 1123
              amount: 10
            type: object
            properties:
              id:
                description: 'Unique id'
                type: integer
              amount:
                description: 'Amount in recipe'
                type: integer
            required:
              - id
              - amount
        tags:
          description: 'List of id tags'
          type: array
          example: [ 1, 2 ]
          items:
            type: integer
        image:
          description: 'Image in Base64'
          example: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABAgMAAABieywaAAAACVBMVEUAAAD///9fX1/S0ecCAAAACXBIWXMAAA7EAAAOxAGVKw4bAAAACklEQVQImWNoAAAAggCByxOyYQAAAABJRU5ErkJggg=='
          type: string
          format: binary
        name:
          description: 'Title'
          type: string
          maxLength: 200
        text:
          description: 'Description'
          type: string
        cooking_time:
          description: 'Cooking time (in minutes)'
          type: integer
          minimum: 1
      required:
        - ingredients
        - tags
        - image
        - name
        - text
        - cooking_time
    ValidationError:
      description: Standard errors in DRG format
      type: object
      properties:
        field_name:
          description: 'Title of field with error. Can be several'
          example: [ 'Mandatory field.' ]
          type: array
          items:
            type: string
    NestedValidationError:
      description: Standard errors in DRG format
      type: object
      properties:
        ingredients:
          description: 'Error in ingredients. In the example above, there were no errors in the first and third ingredient (amount >= 1), but there were in the second.'
          example: [ { },{ "amount": [ "Make sure this value is greater than or equal to 1." ] }, { } ]
          type: array
          items:
            type: object
            properties:
              amount:
                type: array
                items:
                  type: string
...
    AuthenticationError:
      description: User not authorized
      type: object
      properties:
        detail:
          description: 'Error description'
          example: "Credentials were not provided."
          type: string
    PermissionDenied:
      description: Not enough rights
      type: object
      properties:
        detail:
          description: 'Error description'
          example: "You do not have sufficient rights to perform this action."
          type: string
    NotFound:
      description: Object not found
      type: object
      properties:
        detail:
          description: 'Error description'
          example: "Page not found."
          type: string
  responses:
...
    NestedValidationError:
      description: 'Validation errors in the standard DRF format, including those with internal elements.'
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/NestedValidationError'
              - $ref: '#/components/schemas/ValidationError'
    AuthenticationError:
      description: User not authorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthenticationError'
    PermissionDenied:
      description: Not enough rights
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PermissionDenied'
    NotFound:
      description: Object not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFound'
  securitySchemes:
    Token:
      description: 'Auth by token. <br>
      All requests on behalf of the user must be made with the header "Authorization: Token TOKENVALUE"'
      type: http
      scheme: token